+++ platform/darwin/src/MGLStyle.mm
@@ -77,7 +77,7 @@ const MGLExceptionName MGLRedundantSourceIdentifierException = @"MGLRedundantSou
 
 @interface MGLStyle()
 
-@property (nonatomic, readonly, weak) MGLMapView *mapView;
+@property (nonatomic, readonly, weak) id <MGLStylable> stylable;
 @property (nonatomic, readonly) mbgl::style::Style *rawStyle;
 @property (readonly, copy, nullable) NSURL *URL;
 @property (nonatomic, readwrite, strong) NSMutableDictionary<NSString *, MGLOpenGLStyleLayer *> *openGLLayers;
@@ -119,14 +119,14 @@ static_assert(6 == mbgl::util::default_styles::numOrderedStyles,
 
 #pragma mark -
 
-- (instancetype)initWithRawStyle:(mbgl::style::Style *)rawStyle mapView:(MGLMapView *)mapView {
-    MGLLogInfo(@"Initializing %@ with mapView: %@", NSStringFromClass([self class]), mapView);
+- (instancetype)initWithRawStyle:(mbgl::style::Style *)rawStyle stylable:(id <MGLStylable>)stylable {
+    MGLLogInfo(@"Initializing %@ with stylable: %@", NSStringFromClass([self class]), stylable);
     if (self = [super init]) {
-        _mapView = mapView;
+        _stylable = stylable;
         _rawStyle = rawStyle;
         _openGLLayers = [NSMutableDictionary dictionary];
         _localizedLayersByIdentifier = [NSMutableDictionary dictionary];
-        MGLLogDebug(@"Initializing with style name: %@ mapView: %@", self.name, mapView);
+        MGLLogDebug(@"Initializing with style name: %@ stylable: %@", self.name, stylable);
     }
     return self;
 }
@@ -186,17 +186,17 @@ static_assert(6 == mbgl::util::default_styles::numOrderedStyles,
     // TODO: Fill in options specific to the respective source classes
     // https://github.com/mapbox/mapbox-gl-native/issues/6584
     if (auto vectorSource = rawSource->as<mbgl::style::VectorSource>()) {
-        return [[MGLVectorTileSource alloc] initWithRawSource:vectorSource mapView:self.mapView];
+        return [[MGLVectorTileSource alloc] initWithRawSource:vectorSource stylable:self.stylable];
     } else if (auto geoJSONSource = rawSource->as<mbgl::style::GeoJSONSource>()) {
-        return [[MGLShapeSource alloc] initWithRawSource:geoJSONSource mapView:self.mapView];
+        return [[MGLShapeSource alloc] initWithRawSource:geoJSONSource stylable:self.stylable];
     } else if (auto rasterSource = rawSource->as<mbgl::style::RasterSource>()) {
-        return [[MGLRasterTileSource alloc] initWithRawSource:rasterSource mapView:self.mapView];
+        return [[MGLRasterTileSource alloc] initWithRawSource:rasterSource stylable:self.stylable];
     } else if (auto rasterDEMSource = rawSource->as<mbgl::style::RasterDEMSource>()) {
-        return [[MGLRasterDEMSource alloc] initWithRawSource:rasterDEMSource mapView:self.mapView];
+        return [[MGLRasterDEMSource alloc] initWithRawSource:rasterDEMSource stylable:self.stylable];
     } else if (auto imageSource = rawSource->as<mbgl::style::ImageSource>()) {
-        return [[MGLImageSource alloc] initWithRawSource:imageSource mapView:self.mapView];
+        return [[MGLImageSource alloc] initWithRawSource:imageSource stylable:self.stylable];
     } else {
-        return [[MGLSource alloc] initWithRawSource:rawSource mapView:self.mapView];
+        return [[MGLSource alloc] initWithRawSource:rawSource stylable:self.stylable];
     }
 }
 
@@ -211,7 +211,7 @@ static_assert(6 == mbgl::util::default_styles::numOrderedStyles,
     }
 
     try {
-        [source addToMapView:self.mapView];
+        [source addToStylable:self.stylable];
     } catch (std::runtime_error & err) {
         [NSException raise:MGLRedundantSourceIdentifierException format:@"%s", err.what()];
     }
@@ -243,7 +243,7 @@ static_assert(6 == mbgl::util::default_styles::numOrderedStyles,
         }
     }
     
-    return [source removeFromMapView:self.mapView error:outError];
+    return [source removeFromStylable:self.stylable error:outError];
 }